move extension schema
