--- src/main/detector/replayDetectorExtension.js
@@ -1,7 +1,0 @@
-import { BaseExtension } from "../extensions/baseExtension";
-
-export class ReplayDetectorExtension extends BaseExtension {
-  register(listener) {
-    throw new Error("Unimplemented");
-  }
-}
