--- src/main/uploader/replayUploader.js
@@ -1,7 +1,0 @@
-import { BaseExtension } from "../extensions/baseExtension";
-
-export class ReplayUploaderExtension extends BaseExtension {
-  upload(replayData) {
-    throw new Error("Unimplemented");
-  }
-}
